Implement rudamentary backup list support.
This change implements the very beginnings of backup listing support.  Included
with this is the ability to read backup descriptor 2, as well as previous
instances of the descriptor to get a unified picture of the backup set.

I've also added lots of logging and fixed a few bugs I found working with the
file format.

This change includes implementation of StatusOr<T>, something I picked up from
Google -- it's very questionable whether I can actually use this...

There's still a lot of work to do in testing this and cleaning things up.  In
particular, the malloc/free instances used in here are really icky and need to
be cleaned up.