Implement initial full backup restore support.
This change adds support for performing a restoration from a complete backup
set.  For right now it only uses the most recent backup set to do the
restoration, and you can't specify a different one.  It also doesn't understand
things like incremental backups, or sorting the backup chunks to optimize
performance.  It does however support compressed and non-compressed backups, as
well as re-duplicating files that were de-duplicated.

I also did some cleaning up:
 - Moved MD5 and compression calculations into BackupVolume, done when writing
   chunks to the files, and undone when reading the chunks.
 - Added -O3 optimization which greatly speeds up the processing of files (only
   on non-debug builds).
 - Added some comments and moved a bit of code around.

I had to disable some code in the backup volume test to get it to compile like
this -- TODO on the tests.  Also, the backup and restore drivers are not
testible as they stand -- they'll need a good refactoring to get them there.