Make this compile on Windows.
The tests fail and there's a LOT of warnings thrown, but at least it all builds.
The key to getting it all working was to build all dependent libraries with /MD
and copy in the built DLL files when running.  That's the only way I can think
of to replicate what I did to get this working.