Add preliminary multiple-file support.
This change adds the beginnings of multiple-file support to the program,
augmenting the LoadFileSets() function to be able to request additional
BackupVolumes as needed to complete its processing.  The requests are made via
callbacks that the backup driver passes in -- the backup driver can then do
things like prompt the user to insert media, or other such things.

In writing the files, the backup driver need simply close out its backup volume
when its done writing and open a new one to continue going.  As long as the
volume pointers are correct, all should be well.  (This remains to be tested
though.)