Add a test for dealing with binary files, and correctly deal with them.
In Windows, a flag needs to be added to the mode to cause files to be treated as
binary files.  Otherwise, Windows does transformations on the content.  This is
what was causing the backups to fail, though the tests passed.

I've also modified the backup driver to strip leading '/' or drive letters off
paths as they're added to the backup, causing restores to work in Windows.

At this point, both backup and restore work perfectly in Windows, including
compression.