Add functional multi-file support.
This change fills out the multi-file support, adding a BackupLibrary to
coordinate the whole thing.  This simplified things quite a bit and made it a
bit easier to manage the backup volumes.  It also decentralized a lot of stuff
from the volumes, instead putting them in the library.

With this change, each new backup creates a new backup file in numeric order
from the previous, and restores are able to track chunks back to previous
volumes.

I had to disable the backup volume test, as there were too many changes made to
keep the test running at the moment.  Tests will be needed to continue further.