Interface out things in prep for testing.
This change creates interfaces for a bunch of things in preparation to test
BackupLibrary.  Originally it was implemented with a lot of static usage of
objects, which overall makes unit testing impossible.  With this change, it
accepts all dependencies in the construtor as factories or as interfaces, and
the various drivers do the actual creation of objects.   This will allow tests
to implement fakes and mocks where needed.

I also added a bunch of missing comments.