Make FindBasenameAndLastVolume also return the number of volumes.
It turns out only knowing that the last volume is zero doesn't tell you if there
are in fact any volumes.  This change modifies the File API to provide that
also, and plumbs it through BackupLibrary, which simplifies the logic a bit.
I've also modified CreateBackup() to return Status instead of fataling out.