Lots of tweaks to the graphical interface.
Added a progress bar, tweaked the coloring, and moved things around to make the
layout more flexible to changes.  The tab bars are now done with two independent
linked tab widgets, one with just the bar shown, and the other with just the
pane shown.  When one is clicked, the other is changed, so it all acts as one,
but it allowed me to take the tab bar out of document mode and put it in a
separate pane, allowing me to add other things to the left side.