Conver the UI into more of a wizard-style.
This change removes the big tabs in the backup action with a single Backup tab,
and makes the flow more like a wizard.  I've also added a neat progress
indicator to show how many steps are left.

All this accomplished with a whole bunch of tabsets with hidden tab bars that
are all connected together.