Add some directory-management functions to File, GetLabels returns const Label.
This change adds IsDirectory() and ListDirectory() methods to File to deal with
file enumeration.  It also changes the GetLabels() function to return a vector
of Label copies, rather than pointers, so the labels can be used without
requiring the source BackupLibrary to still be around, as will very often
happen when querying for labels.