Fix label naming and loss of labels.
It was the case that creating a backup with a new label would eliminate all
existing labels, because we weren't keeping track of past labels in the library.
This change adds support for that, and also fixes a rename bug in the default
label.

While here, I removed the overwrite question when choosing a destination, and
also removed the use of Label* in all the maps, making portability a little
easier.