Add label history support.
This change adds a new dialog box for showing label history.  With that, I've
added a bunch of extra metadata to backup filesets for user friendliness, and
added functionality to the library and the volume to track down filesets along
lineage lines.  This, plus a simple change in the backup driver, enables us to
create and maintain separated backups in different labels, but de-duping against
the entire pool of data.