Add backup cancelling support to library, partial GUI suport.
This change adds cancelling suport to the backup system, causing it to write a
special bit in the backup descriptor header indicating such.  Operations like
getting labels then look for the last successful backup volume to retreive the
data.

At some point, the extra code in here that supports appending backup sets to
existing files should be removed.  It was originally implemented when I thought
we would be appending to existing files, rather than creating new ones.  But
that design opens a huge can of worms I don't want to deal with, and this
cancelling support breaks that anyway.  So we should just remove that code.