Max size of first volume depends on sizes of last volumes.
This change adds functionality to crop the size of the first new volume in a
backup to account for smaller than max-size volumes before it.  This way, the
combination of that and all smaller volumes before will fit within the max size
(and allows for completely filling up, say, DVD media).