Add elapsed / remaining time, more backup logging, enable or disable VSS.
This change adds elapsed and remaining time estimation to the backup UI, and a
bit more logging during the backup so the user is more aware of what's going on.
I also added a new checkbox to the options to allow a user to enable or disable
VSS when supported.  This way I can test backups without having to run as
administrator.