Rework the restore selector model to work well in linux.
This change required a complete rewrite of the restore selector model.
Originally implemented as a QStandardItemModel, this was actually very slow for
very large sets, and the threading model I was using turned out to not work in
Linux at all.

Instead, I've reimplemented it from scratch as a QAbstractItemModel, with its
own data store in a tree.  This turns out to be much much faster for the massive
amounts of data, as well as insertions and removals.

The threading model I was using to populate the model had to be jettisoned in
favor of manually pumping the event loop.  GUI actions, including inserting and
removing data from a model, need to be done in the main thread, so to keep a
please wait dialog box working through that, it's necessary to interrupt
processing every now and again.  To ensure the processing happens as quickly as
possible, we only pump the event loop for 1ms every 100ms.

Reworking the tree model required use of a test suite to ensure the model is
correct.  I've included the sources for that test here as well to allow for
testing in the future if need be.

At this point, the restore UI is nearing completion, with the selector able to
deal with history and checking/unchecking of items.

We're still excluding empty directories from the list -- that needs to be fixed
before we can realistically use this.