Add back support for empty directories.
This change makes it possible for empty directories to exist in the restore list
again by making the removal of entries in the list more intelligent.  When a
directory is requested for removal in the list, all children are removed as
well, and the leaf list is updated correctly.  Thus, no more crashing.