Implement restore.
This fills in the last remaining functionality in the backup program --
restoring files to their destinations.  With this change, files selected from a
label and a specific time in history can be restored to any directory -- the
directory layout is recreated in that directory.  All UI operations work,
including canceling backups.

There are still some bugs, but predominantly it all works!