Add symlink support to the backup system.
This change introduces a file format change -- older backup files will not work
with this change.

With this change, symlinks are now properly backed up and restored, rather than
followed (at least in Linux).  This change now allows me to completely back up
and restore my home directory, resulting in a near-perfect replication of the
data.  We're still missing file attributes, and modification times are not
restored.