Fix a problem in buffering, add buffering to backup process.
This change fixes a problem in file buffering where I forgot to double the size
of the file buffer.

I've also added changes to allow the chunk-writing to take advantage of the
write buffering.  In practicality, I'm not sure this has increased performance
any, though I seem to recall a backup of Tess's data taking over 24 hours (in
estimate) before, where now it reported 14.  That was pre- her harddrive dying
though.

Finally, I added a change in the backup driver to read in up to 5mb chunks from
the harddrive, but then process only 64kb chunks at a time (with write buffering
back out).